gradesbot/requirements.txt
telethon>=1.24.0
pyrogram>=2.0.0
tgcrypto>=1.2.3
pandas>=2.0.0
polars>=0.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
supabase>=1.0.0
flask>=2.3.0
flask-login>=0.6.0
werkzeug>=2.3.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
requests>=2.31.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
scipy>=1.10.0
cryptography>=41.0.0
